This may have been the root of the issue with "GetGlobalTermWeightFromString" I ran into awhile back - changed the variable that is returned from an int to a float.

+++ trunk/src/main/java/pitt/search/semanticvectors/LuceneUtils.java
@@ -212,7 +212,7 @@ public class LuceneUtils {
    * in all contents fields.
    */
   public float getGlobalTermWeightFromString(String termString) {
-    int freq = 0;
+    float freq = 0;
     for (String field: flagConfig.contentsfields())
       freq += getGlobalTermWeight(new Term(field, termString));
     return freq;
